% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{aggregate_loop}
\alias{aggregate_loop}
\title{Aggregate loop}
\usage{
aggregate_loop(listaShrt, cdat_greenLt, years, type)
}
\arguments{
\item{listaShrt, cdat_greenLt}{A data frame}

\item{years}{A vector of integer}

\item{type}{A character with values "L" when running launch_green and "N"
when running nut_balance}
}
\value{
A matrix
}
\description{
This function aggregate the variables of data frame.
}
\keyword{internal}
