% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{calculate_shreve}
\alias{calculate_shreve}
\title{Calculate shreve}
\usage{
calculate_shreve(catch_data)
}
\arguments{
\item{catch_data}{data frame.}
}
\value{
A data frame
}
\description{
Calculate shreve
}
\keyword{internal}
