% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{calib_green_help}
\alias{calib_green_help}
\title{Parallel calibration help function}
\usage{
calib_green_help(task, catch_data, annual_data, years, latin_range)
}
\arguments{
\item{task}{An integer with the number of simulation}

\item{catch_data}{data frame. It defines the topological sequence
of catchments.}

\item{annual_data}{data frame. It defines the sources of nutrient for each
year and catchments.}

\item{years}{integer. It defines the years of interest.}

\item{latin_range}{data frame. It defines the values for the params taking +
into account the limits.}
}
\value{
A matrix
}
\description{
This is a help function for calib_green function
}
\keyword{internal}
