% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_data_TN.R
\docType{data}
\name{catch_data_TN}
\alias{catch_data_TN}
\title{Catch data TN}
\format{
A data frame with 5 variables:
\describe{
  \item{\code{HydroID}}{integer positive. Unique catchment identifier.}
  \item{\code{To_catch}}{integer. Unique identifier of the catchment to which
  the catchment goes. Note that for the outlet To_catch== -1.}
  \item{\code{Shreve}}{integer. this indicates the Shreve order of the
  topological sequence in the stream network.}
  \item{\code{LakeFrRet}}{fraction, 0-1. Lake retention fraction.}
  \item{\code{NrmLengthKm}}{double. Normalized length of catchment reach.}
}
}
\usage{
catch_data_TN
}
\description{
Defines the topological sequence of catchments for nitrogen.
}
\keyword{datasets}
