% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{check_colnames_annual}
\alias{check_colnames_annual}
\title{Check colnames}
\usage{
check_colnames_annual(df)
}
\arguments{
\item{df}{data frame.}
}
\value{
A data frame
}
\description{
This function checks that all column names for the catch_data
data frame are correct.
}
\keyword{internal}
