% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{check_colnames_catch}
\alias{check_colnames_catch}
\title{Check colnames catch_data}
\usage{
check_colnames_catch(df)
}
\arguments{
\item{df}{data frame.}
}
\value{
A data frame.
}
\description{
This function checks that all column names for the annual_data
data frame are correct.
}
\keyword{internal}
