% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{check_years}
\alias{check_years}
\title{Check years}
\usage{
check_years(loc_years, years)
}
\arguments{
\item{loc_years}{numeric. A vector of integer with the years of interest}

\item{years}{numeric. A vector of integer with the years with information in
data frame}
}
\value{
A vector of integer
}
\description{
This function check if years are in data frame.
}
\keyword{internal}
