% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{create_levels}
\alias{create_levels}
\title{Create levels}
\usage{
create_levels(df, no_sf)
}
\arguments{
\item{df, no_sf}{data frame.}
}
\value{
A list
}
\description{
This function create the levels for the data frame.
}
\keyword{internal}
