% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{create_lits_of_maps}
\alias{create_lits_of_maps}
\title{Preprocessing data for scenario summary}
\usage{
create_lits_of_maps(
  hydroSf,
  long_basin,
  unit,
  legend_position,
  param_list,
  style
)
}
\arguments{
\item{hydroSf}{data frame.}

\item{long_basin}{numeric. The longitude of the basin.}

\item{unit}{character. The unit of the data.}

\item{legend_position}{numeric. It indicates de position of the legend of the
plot. (Default: 1)}

\item{param_list}{list. A list with the maps to plot.}

\item{style}{character. The style of the plot.}
}
\description{
This function blah, blah, blah....
}
\keyword{internal}
