% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{create_map}
\alias{create_map}
\title{Create map}
\usage{
create_map(
  hydroSf,
  var_name,
  style_map,
  scale_barRefs,
  scale_barTexts,
  title,
  palette,
  legend_position = 1
)
}
\arguments{
\item{hydroSf}{data frame.}

\item{var_name}{character. The name of a data variable that it is contains
in data frame}

\item{style_map, scale_barRefs, scale_barTexts}{numeric.}

\item{title}{character. The title of the plot}

\item{palette}{character. The colors of the map}

\item{legend_position}{numeric. It indicates de position of the legend of the
plot. (Default: 1)}
}
\value{
A plot with map of the basin
}
\description{
This function create a tm map.
}
\keyword{internal}
