% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{data_preparation}
\alias{data_preparation}
\title{data_preparation}
\usage{
data_preparation(catch_data, annual_data)
}
\arguments{
\item{catch_data}{data frame. Definition of the topological sequence of
catchments.}

\item{annual_data}{data frame. Sources of nutrient for each year and
catchments.}
}
\value{
One list with two data table
}
\description{
Prepare data to launch green
}
\keyword{internal}
