% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{evolution_plot}
\alias{evolution_plot}
\title{Evolution plot}
\usage{
evolution_plot(
  data,
  title = NULL,
  xaxis.title = NULL,
  yaxis.title,
  x,
  y,
  colour = NULL,
  wrap = NULL
)
}
\arguments{
\item{data}{data frame.}

\item{title}{character. The plot title}

\item{xaxis.title}{character. The title of x axis}

\item{yaxis.title}{character. The title of y axis}

\item{x}{character. The name of variable in data to represent in x axis}

\item{y}{character. The name of variable in data to represent in y axis}

\item{colour}{character. The name of variable in data to represent colour}

\item{wrap}{character. The name of variable in data to represent wrap}
}
\description{
This function blah, blah, blah....
}
\keyword{internal}
