% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{evolution_plot_area}
\alias{evolution_plot_area}
\title{Evolution plot area}
\usage{
evolution_plot_area(
  annual_data,
  new_data,
  title_plot = NULL,
  xaxis.title = NULL,
  yaxis.title,
  x,
  y,
  colour = NULL,
  wrap = NULL
)
}
\arguments{
\item{annual_data, new_data}{data frame.}

\item{title_plot}{character. The plot title}

\item{xaxis.title}{character. The title of x axis}

\item{yaxis.title}{character. The title of y axis}

\item{x}{character. The name of variable in data to represent in x axis}

\item{y}{character. The name of variable in data to represent in y axis}

\item{colour}{character. The name of variable in data to represent colour}

\item{wrap}{character. The name of variable in data to represent wrap}
}
\value{
A plot
}
\description{
This function blah, blah, blah....
}
\keyword{internal}
