% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_metrics.R
\name{gof}
\alias{gof}
\title{GOF}
\usage{
gof(sim, obs, na.rm = TRUE, digits = 2)
}
\arguments{
\item{sim}{numeric 'data.frame', 'matrix' or 'vector' with simulated values}

\item{obs}{numeric 'data.frame', 'matrix' or 'vector' with observed values}

\item{na.rm}{Should remove NA}

\item{digits}{Number of digits}
}
\value{
A data frame
}
\description{
GOF
}
\keyword{internal}
