% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{LakeRetent_plot}
\alias{LakeRetent_plot}
\title{Lake retention values summary}
\usage{
LakeRetent_plot(catch_data_TN)
}
\arguments{
\item{catch_data_TN}{data frame. Sources of nutrient for each year and
catchments.}
}
\value{
barplot & histogram-density
}
\description{
Summary of the reference values in the stations
}
\examples{
\donttest{
# the data of the TN scenario
data(catch_data_TN)
LakeRetent_plot(catch_data_TN)
}

}
