% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{load_map}
\alias{load_map}
\title{Load DA map}
\usage{
load_map(hydroSf, long_basin, title, style_map, legend_position = 1)
}
\arguments{
\item{hydroSf}{data frame.}

\item{long_basin}{numeric.}

\item{title}{string. The title of the map}

\item{style_map}{character. The style of the map}

\item{legend_position}{numeric. It indicates de position of the legend of the
plot. (Default: 1)}
}
\value{
A plot with map of the basin
}
\description{
This function creates a tmap
}
\keyword{internal}
