% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{load_SA_map}
\alias{load_SA_map}
\title{Load SA map}
\usage{
load_SA_map(
  hydroSf_merge,
  refN_P,
  long_basin,
  unity,
  style_map,
  legend_position = 1
)
}
\arguments{
\item{hydroSf_merge}{data frame.}

\item{refN_P}{numeric. The number of variables of the data frame}

\item{long_basin}{numeric.}

\item{unity}{character. The units of hydroSf}

\item{style_map}{character. The style of the map}

\item{legend_position}{numeric. It indicates de position of the legend of the
plot. (Default: 1)}
}
\value{
A plot with map of the basin
}
\description{
This function creates a tmap
}
\keyword{internal}
