% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{multiple_map}
\alias{multiple_map}
\title{Multiple Map}
\usage{
multiple_map(hydroSf, refN_P, long_basin, unit, legend_position = 1)
}
\arguments{
\item{hydroSf}{data frame.}

\item{refN_P}{numeric. The number of variables of the data frame}

\item{long_basin}{numeric.}

\item{unit}{character. The units of hydroSf}

\item{legend_position}{numeric. It indicates de position of the legend of the
plot. (Default: 1)}
}
\value{
A plot with map of the basin
}
\description{
This function creates an arrange of tmap
}
\keyword{internal}
