% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{nutrient_maps}
\alias{nutrient_maps}
\title{Map average load output by source}
\usage{
nutrient_maps(green_file, sh_file, basin_name, plot.type, legend_position = 1)
}
\arguments{
\item{green_file}{data frame of GREEN model results from green_shares()
function. Nutrient Load by source apportionment of nutrient for each year
and catchments.}

\item{sh_file}{sf object. The spatial information of the basin.}

\item{basin_name}{character. The title of the map.}

\item{plot.type}{character. Alternatives of the map: “gr1”: output load
(kt/y) by source; “gr2”: Total Load, log10 (kt/y); “gr3”: Total Load
by km2 (kt/year/km2).}

\item{legend_position}{numeric. Legend position: 1 (default): "right",
"bottom"; 2: "left", "up"; 3: "right", "bottom"; 4: "right", "up".}
}
\value{
A map
}
\description{
Creates maps showing basin output total or by source loads
}
\examples{
\donttest{
# the data of the TN scenario
data(catch_data_TN)
data(annual_data_TN)
data(sh_file)
# the parameter to assess the basin model
alpha_p <- 35.09
alpha_l <- 0.02
sd_coef <- 0.2
# years in which the model should be executed
loc_years <- 1990:2018
# Computing the source apportionment
basin_sa <- green_shares(catch_data_TN, annual_data_TN, alpha_p, alpha_l,
sd_coef, loc_years)
# The title of the Map
mapTitle <- "Output Loads  for the Lay Basin"
# Basin Output Load  Maps by source
Lpos <- 1
nutrient_maps(basin_sa, sh_file, mapTitle, "gr1", Lpos)
# Basin Output Specific Load  Maps
Lpos <- 1
nutrient_maps(basin_sa, sh_file, mapTitle, "gr2", Lpos)
# Basin Output Specific Load by km2 Maps
Lpos <- 1
nutrient_maps(basin_sa, sh_file, mapTitle, "gr3", Lpos)
}

}
