% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{preproc_scenSummary}
\alias{preproc_scenSummary}
\title{Preprocessing data for scenario summary}
\usage{
preproc_scenSummary(annual_data)
}
\arguments{
\item{annual_data}{data frame.}
}
\value{
A data frame with the total load by year and type
}
\description{
This function blah, blah, blah....
}
\keyword{internal}
