% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{references_plot}
\alias{references_plot}
\title{Reference summary plot}
\usage{
references_plot(annual_data)
}
\arguments{
\item{annual_data}{data frame. Sources of nutrient for each year and
catchments.}
}
\value{
A barplot, a histogram-density and a boxplot
}
\description{
Summary of the reference values in the stations
}
\examples{
\donttest{
# the data of the TN scenario
data(annual_data_TN)
references_plot(annual_data_TN)
}
}
