% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{shreve}
\alias{shreve}
\title{Shreve}
\usage{
shreve(the_SC)
}
\arguments{
\item{the_SC}{table. A table with topology data.}
}
\value{
One object, a data frame with the shreve.
}
\description{
Function to read the data and return the data frame for GREEN
execution.
}
