% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{simobs_annual_plot}
\alias{simobs_annual_plot}
\title{Facet year plot}
\usage{
simobs_annual_plot(
  catch_data,
  annual_data,
  alpha_p,
  alpha_l,
  sd_coef,
  years,
  name_basin,
  maxvalue
)
}
\arguments{
\item{catch_data}{data frame. Definition of the topological sequence of
catchments.}

\item{annual_data}{data frame. Sources of nutrient for each year and
catchments.}

\item{alpha_p}{numeric. First model parameter, the basin retention
coefficient.}

\item{alpha_l}{numeric. Second model parameter, the river retention
coefficient.}

\item{sd_coef}{numeric. Third model parameter, fraction of domestic diffuse
sources that reaches the stream network.}

\item{years}{integer. Years to be used in the calibration. For sequences use
c(yearini:yearend).}

\item{name_basin}{character. The name of the basin}

\item{maxvalue}{numeric. The maximum value}
}
\value{
One object, a data frame
}
\description{
This function blah, blah, blah....
}
