% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen_plots.R
\name{gr_density_plot}
\alias{gr_density_plot}
\title{Density plot}
\usage{
gr_density_plot(df_plot, plot_index, basin_name, cSD)
}
\arguments{
\item{df_plot}{data frame.}

\item{plot_index}{numeric. The indexes to plot}

\item{basin_name}{character. The name of the basin}

\item{cSD}{numeric. The standard deviation}
}
\value{
No return value, called for the side effect of drawing a plot
}
\description{
This function blah, blah, blah....
}
\keyword{internal}
