% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgreen.R
\name{read_geometry}
\alias{read_geometry}
\title{Read geometry}
\usage{
read_geometry(path, name)
}
\arguments{
\item{path}{string. A string with the path of the shp file.}

\item{name}{string. A string with the name of the geometry file.}
}
\value{
One object, a sf file.
}
\description{
Function to read the geometry file.
}
